Split out OpenAuthoringAuthModal to its own file
